+++ lib/CwpException.php
@@ -6,7 +6,7 @@
  * for the client area.
  *
  * @package cwp7
- * @version 2.0.0
+ * @version 2.0.1
  * @author  Booysen Logistics <bradley@booysenlogistics.co.za>
  * @license MIT
  * @link    https://github.com/bradleygb/whmcs-cwp-module